Templates moved to a separate package.
